backend/requirements-prod.txt
# Production dependencies
gunicorn==21.2.0  # WSGI server for production
psycopg2-binary==2.9.9  # PostgreSQL adapter
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
email-validator==2.1.0
redis==5.0.1  # For caching and session management
celery==5.3.4  # For background tasks
